requirements.txt
# Core Flask dependencies
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# Flask extensions
flask-cors==6.0.1

# Application dependencies
wakeonlan==3.1.0
python-dotenv==1.2.1
